Milestone 1
The article page is now displays a comprehensive list of all articles available in the system. When a user decides to create a new article, we take immediate action by saving the test article name in the article table, and then redirecting the user to the article edit page.

Additionally, on the article home page, we've implemented an "Edit" button. When clicked, this button guides the user to the article edit page.

To implement this, we've introduced an "ArticleObject" class and a corresponding "ArticleDTO". Alongside these additions, a new controller, "ArticleController," has been incorporated into the system. This controller's primary responsibility is to manage article creation and edit requests, ensuring they are correctly routed to the corresponding methods within the entity repository.

This process begins with the creation of articles through an AJAX call, enabling us to instantly redirect the user to the edit page.