Milestone 1
To implement article editing functionality, we have introduced a two-panel layout on the article editing page.

The first panel displays the TinyMCE editor, where we have also included its comments plugin.

The second panel contains required tabs. Currently, we have populated the "Fields" tab with a dropdown containing predefined values. In the future, we plan to integrate FormioJS to facilitate the creation of additional necessary fields.