Milestone 1
We have implemented the feature that allows users to write an articles using the TinyMCE editor on the article editing page.

Additionally, we have implemented an auto save function, that currently storing the article content directly in the article table. Whereas, in the future, we plan to incorporate with an "article_temp" table.

We have also given users the ability to manually save their articles at any point during the writing process.

We have introduced the required tabs on the second panel. Soon, we'll be integrating formio under the "fields" tab for required functionality.