Milestone 1
We have introduced a new feature that allows users to add comments to articles. To facilitate this, we've added a menu item that lets users toggle the visibility of the comments sidebar.

When the comments sidebar is active, users have the ability to both add and edit comments within a conversation.

Additionally, we will incorporate more features into the comments plugin soon i.e. delete or reply on a comments, delete a conversation etc.